refactor: screen displaying
